🔨 Refactor :  card , theme 리팩토링  type 공통화
+type 공통화
